+++ test/fixtures/bigv1objects.mbtiles
