+++ validators/bigV1Objects/README.md
@@ -1,0 +1,8 @@
+#### Description
+
+This validator detects all the large new objects created that has version 1 in the last 3 days
+
+#### Usage
+
+1. Download [OSM QA Tiles](https://osmlab.github.io/osm-qa-tiles/) for the planet or a country that you are interested in. 
+2. For example, to run for peru.mbtiles: `osmlint bigv1objects peru.mbtiles > peru.json`
