+++ validators/bigV1Objects/map.js
@@ -1,19 +1,25 @@
 'use strict';
+var time = require('time')(Date);
 var turf = require('turf');
+var today = (time.time() - 3 * 24 * 60 * 60);
 
 module.exports = function(tileLayers, tile, writeData, done) {
   var layer = tileLayers.osm.osm;
-  var osmlint = 'waterbodiesv1';
+  var osmlint = 'bigv1objects';
   var result = [];
   for (var i = 0; i < layer.features.length; i++) {
     var val = layer.features[i];
-    if (val.properties.natural && val.properties.natural === 'water' && val.geometry.type === 'Polygon' && val.properties['@version'] === 1 && val.properties['@timestamp'] > '1464739200') {
+    if ((val.properties.aeroway ||
+        val.properties.leisure ||
+        val.properties.landuse ||
+        val.properties.man_made ||
+        val.properties.building) &&
+      val.geometry.type === 'Polygon' && val.properties['@version'] === 1 && val.properties['@timestamp'] > today) {
       var fc = {
         type: 'FeatureCollection',
         'features': [val]
       };
       var area = turf.area(fc);
-
       if (area > 200000) {
         val.properties._osmlint = osmlint;
         result.push(val);
@@ -22,8 +28,8 @@ module.exports = function(tileLayers, tile, writeData, done) {
   }
 
   if (result.length > 0) {
-    //var fc = turf.featurecollection(result);
-    writeData(JSON.stringify(fc) + '\n');
+    var features = turf.featureCollection(result);
+    writeData(JSON.stringify(features) + '\n');
   }
 
   done(null, null);
