--- validators/waterBodiesV1/README.md
@@ -1,8 +1,0 @@
-#### Description
-
-This validator detects all the large new water bodies created that has version 1.   
-
-#### Usage
-
-1. Download [OSM QA Tiles](https://osmlab.github.io/osm-qa-tiles/) for the planet or a country that you are interested in. 
-2. For example, to run for peru.mbtiles: `osmlint deprecatehighways peru.mbtiles > peru.json`
